+++ error.go
@@ -33,6 +33,14 @@ import (
 	"go.elastic.co/apm/stacktrace"
 )
 
+const (
+	// maxErrorGraphSize is the maximum number of errors
+	// to report in an error tree. Once this number of
+	// nodes is reached, we will stop recursing through
+	// error causes.
+	maxErrorTreeNodes = 50
+)
+
 // Recovered creates an Error with t.NewError(err), where
 // err is either v (if v implements error), or otherwise
 // fmt.Errorf("%v", v). The value v is expected to have
@@ -85,12 +93,10 @@ func (t *Tracer) NewError(err error) *Error {
 	e.cause = err
 	e.err = err.Error()
 	rand.Read(e.ID[:]) // ignore error, can't do anything about it
-	initException(&e.exception, err)
-	initStacktrace(e, err)
-	if len(e.stacktrace) == 0 {
+	initException(&e.exception, err, e.stackTraceLimit)
+	if len(e.exception.stacktrace) == 0 {
 		e.SetStacktrace(2)
 	}
-	e.exceptionStacktraceFrames = len(e.stacktrace)
 	return e
 }
 
@@ -115,9 +121,7 @@ func (t *Tracer) NewErrorLog(r ErrorLogRecord) *Error {
 	e.err = e.log.Message
 	rand.Read(e.ID[:]) // ignore error, can't do anything about it
 	if r.Error != nil {
-		initException(&e.exception, r.Error)
-		initStacktrace(e, r.Error)
-		e.exceptionStacktraceFrames = len(e.stacktrace)
+		initException(&e.exception, r.Error, e.stackTraceLimit)
 	}
 	return e
 }
@@ -152,8 +156,9 @@ type Error struct {
 	// the error's Send method is called.
 	*ErrorData
 
-	// cause holds original error.
-	// It is accessible by Cause method
+	// cause holds the original error.
+	//
+	// It is accessible via the Cause method:
 	// https://godoc.org/github.com/pkg/errors#Cause
 	cause error
 
@@ -166,16 +171,11 @@ type Error struct {
 type ErrorData struct {
 	tracer             *Tracer
 	stackTraceLimit    int
-	stacktrace         []stacktrace.Frame
 	exception          exceptionData
 	log                ErrorLogRecord
+	logStacktrace      []stacktrace.Frame
 	transactionSampled bool
 	transactionType    string
-
-	// exceptionStacktraceFrames holds the number of stacktrace
-	// frames for the exception; stacktrace may hold frames for
-	// both the exception and the log record.
-	exceptionStacktraceFrames int
 
 	// ID is the unique identifier of the error. This is set by
 	// the various error constructors, and is exposed only so
@@ -332,10 +332,10 @@ func (e *ErrorData) enqueue() {
 
 func (e *ErrorData) reset() {
 	*e = ErrorData{
-		tracer:     e.tracer,
-		stacktrace: e.stacktrace[:0],
-		Context:    e.Context,
-		exception:  e.exception,
+		tracer:        e.tracer,
+		logStacktrace: e.logStacktrace[:0],
+		Context:       e.Context,
+		exception:     e.exception,
 	}
 	e.Context.reset()
 	e.exception.reset()
@@ -343,14 +343,19 @@ func (e *ErrorData) reset() {
 }
 
 type exceptionData struct {
-	message string
+	message    string
+	stacktrace []stacktrace.Frame
+	cause      []exceptionData
 	ErrorDetails
 }
 
 func (e *exceptionData) reset() {
 	*e = exceptionData{
+		cause:      e.cause[:0],
+		stacktrace: e.stacktrace[:0],
 		ErrorDetails: ErrorDetails{
-			attrs: e.attrs,
+			attrs: e.ErrorDetails.attrs,
+			Cause: e.ErrorDetails.Cause[:0],
 		},
 	}
 	for k := range e.attrs {
@@ -358,17 +363,43 @@ func (e *exceptionData) reset() {
 	}
 }
 
-func initException(e *exceptionData, err error) {
+func initException(e *exceptionData, err error, stackTraceLimit int) {
+	b := exceptionDataBuilder{stackTraceLimit: stackTraceLimit}
+	b.init(e, err)
+}
+
+type exceptionDataBuilder struct {
+	stackTraceLimit int
+	errorCount      int
+	pointerErrors   map[uintptr]struct{}
+}
+
+func (b *exceptionDataBuilder) init(e *exceptionData, err error) bool {
+	b.errorCount++
+	reflectValue := reflect.ValueOf(err)
+	reflectType := reflectValue.Type()
+	switch reflectType.Kind() {
+	case reflect.Chan, reflect.Func, reflect.Map, reflect.Ptr, reflect.Slice, reflect.UnsafePointer:
+		// Prevent infinite recursion due to cyclic error causes.
+		ptrVal := reflectValue.Pointer()
+		if b.pointerErrors == nil {
+			b.pointerErrors = map[uintptr]struct{}{ptrVal: struct{}{}}
+		} else {
+			if _, ok := b.pointerErrors[ptrVal]; ok {
+				return false
+			}
+			b.pointerErrors[ptrVal] = struct{}{}
+		}
+	}
+
 	e.message = truncateString(err.Error())
 	if e.message == "" {
 		e.message = "[EMPTY]"
 	}
-	err = errors.Cause(err)
 
-	errType := reflect.TypeOf(err)
-	namedType := errType
-	if errType.Name() == "" && errType.Kind() == reflect.Ptr {
-		namedType = errType.Elem()
+	namedType := reflectType
+	if reflectType.Name() == "" && reflectType.Kind() == reflect.Ptr {
+		namedType = reflectType.Elem()
 	}
 	e.Type.Name = namedType.Name()
 	e.Type.PackagePath = namedType.PkgPath()
@@ -395,8 +426,22 @@ func initException(e *exceptionData, err error) {
 		e.Code.Number = err.Code()
 	}
 
+	// If the error implements an Unwrap or Cause method, use that to set the cause error.
+	// Unwrap is defined by errors wrapped using fmt.Errorf, while Cause is defined by
+	// errors wrapped using pkg/errors.Wrap.
+	switch err := err.(type) {
+	case interface{ Unwrap() error }:
+		if cause := err.Unwrap(); cause != nil {
+			e.ErrorDetails.Cause = append(e.ErrorDetails.Cause, cause)
+		}
+	case interface{ Cause() error }:
+		if cause := err.Cause(); cause != nil {
+			e.ErrorDetails.Cause = append(e.ErrorDetails.Cause, cause)
+		}
+	}
+
 	// Run registered ErrorDetailers over the error.
-	for _, ed := range typeErrorDetailers[errType] {
+	for _, ed := range typeErrorDetailers[reflectType] {
 		ed.ErrorDetails(err, &e.ErrorDetails)
 	}
 	for _, ed := range errorDetailers {
@@ -406,9 +451,21 @@ func initException(e *exceptionData, err error) {
 	e.Code.String = truncateString(e.Code.String)
 	e.Type.Name = truncateString(e.Type.Name)
 	e.Type.PackagePath = truncateString(e.Type.PackagePath)
+	b.initErrorStacktrace(&e.stacktrace, err)
+
+	for _, err := range e.ErrorDetails.Cause {
+		if b.errorCount >= maxErrorTreeNodes {
+			break
+		}
+		var data exceptionData
+		if b.init(&data, err) {
+			e.cause = append(e.cause, data)
+		}
+	}
+	return true
 }
 
-func initStacktrace(e *Error, err error) {
+func (b *exceptionDataBuilder) initErrorStacktrace(out *[]stacktrace.Frame, err error) {
 	type internalStackTracer interface {
 		StackTrace() []stacktrace.Frame
 	}
@@ -418,14 +475,13 @@ func initStacktrace(e *Error, err error) {
 	switch stackTracer := err.(type) {
 	case internalStackTracer:
 		stackTrace := stackTracer.StackTrace()
-		if e.stackTraceLimit >= 0 && len(stackTrace) > e.stackTraceLimit {
-			stackTrace = stackTrace[:e.stackTraceLimit]
+		if b.stackTraceLimit >= 0 && len(stackTrace) > b.stackTraceLimit {
+			stackTrace = stackTrace[:b.stackTraceLimit]
 		}
-		e.stacktrace = append(e.stacktrace[:0], stackTrace...)
+		*out = append(*out, stackTrace...)
 	case errorsStackTracer:
 		stackTrace := stackTracer.StackTrace()
-		e.stacktrace = e.stacktrace[:0]
-		pkgerrorsutil.AppendStacktrace(stackTrace, &e.stacktrace, e.stackTraceLimit)
+		pkgerrorsutil.AppendStacktrace(stackTrace, out, b.stackTraceLimit)
 	}
 }
 
@@ -433,14 +489,11 @@ func initStacktrace(e *Error, err error) {
 // skipping the first skip number of frames, excluding
 // the SetStacktrace function.
 func (e *Error) SetStacktrace(skip int) {
-	retain := e.stacktrace[:0]
-	limit := e.stackTraceLimit
+	out := &e.exception.stacktrace
 	if e.log.Message != "" {
-		// This is a log error; the exception stacktrace
-		// is unaffected by SetStacktrace in this case.
-		retain = e.stacktrace[:e.exceptionStacktraceFrames]
+		out = &e.logStacktrace
 	}
-	e.stacktrace = stacktrace.AppendStacktrace(retain, skip+1, limit)
+	*out = stacktrace.AppendStacktrace((*out)[:0], skip+1, e.stackTraceLimit)
 }
 
 // ErrorLogRecord holds details of an error log record.
@@ -504,21 +557,25 @@ func init() {
 				details.SetAttr("addr", fmt.Sprintf("%s:%s", addr.Network(), addr.String()))
 			}
 		}
+		details.Cause = append(details.Cause, opErr.Err)
 	}))
 	RegisterTypeErrorDetailer(reflect.TypeOf(&os.LinkError{}), ErrorDetailerFunc(func(err error, details *ErrorDetails) {
 		linkErr := err.(*os.LinkError)
 		details.SetAttr("op", linkErr.Op)
 		details.SetAttr("old", linkErr.Old)
 		details.SetAttr("new", linkErr.New)
+		details.Cause = append(details.Cause, linkErr.Err)
 	}))
 	RegisterTypeErrorDetailer(reflect.TypeOf(&os.PathError{}), ErrorDetailerFunc(func(err error, details *ErrorDetails) {
 		pathErr := err.(*os.PathError)
 		details.SetAttr("op", pathErr.Op)
 		details.SetAttr("path", pathErr.Path)
+		details.Cause = append(details.Cause, pathErr.Err)
 	}))
 	RegisterTypeErrorDetailer(reflect.TypeOf(&os.SyscallError{}), ErrorDetailerFunc(func(err error, details *ErrorDetails) {
 		syscallErr := err.(*os.SyscallError)
 		details.SetAttr("syscall", syscallErr.Syscall)
+		details.Cause = append(details.Cause, syscallErr.Err)
 	}))
 	RegisterTypeErrorDetailer(reflect.TypeOf(syscall.Errno(0)), ErrorDetailerFunc(func(err error, details *ErrorDetails) {
 		errno := err.(syscall.Errno)
@@ -612,6 +669,9 @@ type ErrorDetails struct {
 		//     }
 		Number float64
 	}
+
+	// Cause holds the errors that were the cause of this error.
+	Cause []error
 }
 
 // SetAttr sets the attribute with key k to value v.